backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-decouple>=3.8
httpx>=0.24.0,<0.25.0
supabase>=2.0.0
pydantic[email]>=2.4.0
pydantic-settings>=2.0.0
email-validator>=2.0.0 